requirements.txt
asyncio
aiohttp
asyncpg
asyncpg_stubs
cachetools
revolt.py
jishaku
python-dotenv
typing_extensions
